Add pyUSD to Aave V3
