--- test/plugins/individual-collateral/aave-v3/helpers.ts
@@ -1,4 +1,0 @@
-import { getResetFork } from '../helpers'
-import { FORK_BLOCK } from './constants'
-
-export const resetFork = getResetFork(FORK_BLOCK)
